add sweet alert

+++ script.js
@@ -3,7 +3,13 @@ const listContainer = document.querySelector(".list-container");
 const trash = document.querySelector(".trash");
 const addTask = document.querySelector(".add-task");
 
-addTask.addEventListener("click", () => {
+inputBox.addEventListener("keydown", (e) => {
+    if (e.key === "Enter") {
+        addTaskFunction();
+    }
+});
+
+const addTaskFunction = () => {
     if (inputBox.value === "") {
         Swal.fire("Gagal", "kamu belum memasukkan task kamu", "question");
     } else {
@@ -16,21 +22,59 @@ addTask.addEventListener("click", () => {
     }
     saveData();
     inputBox.value = "";
-});
+};
+
+addTask.addEventListener("click", addTaskFunction);
 
 listContainer.addEventListener(
     "click",
     (e) => {
-        if (e.target.tagName === "LI") {
+        if (e.target.classList.contains("checked")) {
+            showAlert("warning", "Task ini sudah kamu kerjakan");
+        } else if (e.target.tagName === "LI") {
             e.target.classList.toggle("checked");
+            showAlert("success", "Kamu berhasil menyelesaikan tugas ini");
             saveData();
         } else if (e.target.tagName === "IMG") {
-            e.target.parentElement.remove();
-            deleteTask();
+            Swal.fire({
+                title: "Apakah kamu yakin?",
+                text: "Kamu akan menghapus selamanya task ini",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Hapus",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire(
+                        "Terhapus!",
+                        "Task kamu berhasil dihapus.",
+                        "success"
+                    );
+                    e.target.parentElement.remove();
+                    deleteTask();
+                }
+            });
         }
     },
     false
 );
+
+const showAlert = (icon, title) => {
+    const Toast = Swal.mixin({
+        toast: true,
+        position: "top-end",
+        showConfirmButton: false,
+        timer: 3000,
+        timerProgressBar: true,
+        didOpen: (toast) => {
+            toast.addEventListener("mouseenter", Swal.stopTimer);
+            toast.addEventListener("mouseleave", Swal.resumeTimer);
+        },
+    });
+
+    Toast.fire({ icon, title });
+};
 
 const saveData = () => {
     localStorage.setItem("task", listContainer.innerHTML);
